Factor out common reporter functionality in parent trait
