+++ silencer-plugin/src/main/scala-2.13+/com/github/ghik/silencer/SuppressingReporter.scala
@@ -1,8 +1,5 @@
 package com.github.ghik.silencer
 
-import java.io.File
-
-import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import scala.reflect.internal.util.{Position, SourceFile}
 import scala.reflect.io.AbstractFile
@@ -14,12 +11,7 @@ class SuppressingReporter(
   globalFilters: List[Regex],
   pathFilters: List[Regex],
   sourceRoots: List[AbstractFile]
-) extends ForwardingReporter(original) {
-
-  private val deferredWarnings = new mutable.HashMap[SourceFile, ArrayBuffer[(Position, String)]]
-  private val fileSuppressions = new mutable.HashMap[SourceFile, List[Suppression]]
-  private val normalizedPathCache = new mutable.HashMap[SourceFile, String]
-
+) extends ForwardingReporter(original) with SuppressingReporterBase {
   //Suppressions are sorted by end offset of their suppression ranges so that nested suppressions come before
   //their containing suppressions. This is ensured by FindSuppressions traverser in SilencerPlugin.
   //This order is important for proper unused @silent annotation detection.
@@ -37,9 +29,6 @@ class SuppressingReporter(
     for ((pos, msg) <- deferredWarnings.remove(source).getOrElse(Seq.empty))
       warning(pos, msg) // will invoke `filter`
   }
-
-  def checkUnused(source: SourceFile): Unit =
-    fileSuppressions(source).foreach(_.reportUnused(this))
 
   override def reset(): Unit = {
     super.reset()
@@ -80,13 +69,4 @@ class SuppressingReporter(
       case n => n
     }
   }
-
-  private def relativize(dir: AbstractFile, child: AbstractFile): Option[String] = {
-    val childPath = child.canonicalPath
-    val dirPath = dir.canonicalPath + File.separator
-    if (childPath.startsWith(dirPath)) Some(childPath.substring(dirPath.length)) else None
-  }
-
-  private def anyMatches(patterns: List[Regex], value: String): Boolean =
-    patterns.exists(_.findFirstIn(value).isDefined)
 }
